+++ Module.php
@@ -9,11 +9,6 @@ class Module extends \humhub\components\Module
     /**
      * @inheritdoc
      */
-    public $resourcesPath = 'resources';
-
-    /**
-     * @inheritdoc
-     */
     public function getConfigUrl()
     {
         return Url::to(['/example-basic/admin']);